[circle-mpqsolver] Revisit MPQSolver
This commit moves all functions, datatypes and variables used solely
in PatternSolver to PatternSolver.

ONE-DCO-1.0-Signed-off-by: s.malakhov <s.malakhov@partner.samsung.com>
