+++ compiler/circle-mpqsolver/src/MPQSolver.cpp
@@ -16,15 +16,10 @@
 
 #include "MPQSolver.h"
 
-#include "pattern/PatternResolver.h"
-
 #include <luci/ImporterEx.h>
 #include <iostream>
 
 using namespace mpqsolver;
-
-using LayerParam = luci::CircleQuantizer::Options::LayerParam;
-using LayerParams = luci::CircleQuantizer::Options::LayerParams;
 
 MPQSolver::MPQSolver(const std::string &input_data_path, float qerror_ratio,
                      const std::string &input_quantization, const std::string &output_quantization)
@@ -37,53 +32,6 @@ MPQSolver::MPQSolver(const std::string &input_data_path, float qerror_ratio,
 void MPQSolver::set_save_intermediate(const std::string &save_path)
 {
   _hooks = std::make_unique<core::DumpingHooks>(save_path);
-}
-
-void MPQSolver::set_mpq_options(MPQOptions &options) { _options = options; }
-
-LayerParams MPQSolver::get_frozen_params() const
-{
-  LayerParams params;
-  for (auto node_to_param : _frozen._node_to_param)
-  {
-    params.push_back(std::make_shared<LayerParam>(node_to_param.second));
-  }
-
-  return params;
-}
-
-void MPQSolver::resolve_patterns(luci::Module *module)
-{
-  _frozen._node_to_param.clear();
-
-  for (auto pattern : _options._patterns)
-  {
-    std::unique_ptr<pattern::PatternResolver> resolver;
-    switch (pattern)
-    {
-      case QuantizationPattern::Q8LayerNormWithQ16Variance:
-        resolver = std::make_unique<pattern::Q8LayerNormWithQ16VarianceResolver>();
-        break;
-      default:
-        throw std::runtime_error("Unsupported pattern to resolve");
-    }
-
-    auto const resolved = resolver->resolve(module);
-    for (auto node_param : resolved)
-    {
-      auto const frozen = _frozen._node_to_param.find(node_param.first);
-      if (frozen == _frozen._node_to_param.end())
-      {
-        // node was not previously defined - just set it (no ambiguity)
-        _frozen._node_to_param[node_param.first] = node_param.second;
-      }
-      else if (frozen->second.dtype != node_param.second.dtype)
-      {
-        // ambiguity (incoming description conflicts with current)
-        throw std::runtime_error("Resolved patterns contradict each other");
-      }
-    }
-  }
 }
 
 std::unique_ptr<luci::Module> MPQSolver::read_module(const std::string &path)
